check indexer state for each pr
